refactor: move Git to VCS abstraction layer (#50)
Refactor Git implementation to use VCS abstraction layer.

- Move GitManager to vcs/git/ and implement VCSManager
- Rename BranchAnalyzer to GitBranchAnalyzer
- Update old git/ module with deprecation warnings
- Re-export from new location for backwards compatibility
