+++ src/vibe_heal/git/__init__.py
@@ -1,12 +1,29 @@
-"""Git integration for vibe-heal."""
+"""DEPRECATED: Use vibe_heal.vcs instead.
 
-from vibe_heal.git.exceptions import (
+Git integration for vibe-heal - now provided by vcs module for backwards compatibility.
+"""
+
+import warnings
+
+from vibe_heal.vcs.exceptions import (
     DirtyWorkingDirectoryError,
-    GitError,
-    GitOperationError,
-    NotAGitRepositoryError,
 )
-from vibe_heal.git.manager import GitManager
+from vibe_heal.vcs.exceptions import (
+    NotARepositoryError as NotAGitRepositoryError,
+)
+from vibe_heal.vcs.exceptions import (
+    VCSError as GitError,
+)
+from vibe_heal.vcs.exceptions import (
+    VCSOperationError as GitOperationError,
+)
+from vibe_heal.vcs.git.manager import GitManager
+
+warnings.warn(
+    "vibe_heal.git is deprecated. Use vibe_heal.vcs instead.",
+    DeprecationWarning,
+    stacklevel=2,
+)
 
 __all__ = [
     "DirtyWorkingDirectoryError",
