+++ src/vibe_heal/vcs/git/__init__.py
@@ -1,0 +1,9 @@
+"""Git VCS implementation for vibe-heal."""
+
+from vibe_heal.vcs.git.branch_analyzer import GitBranchAnalyzer
+from vibe_heal.vcs.git.manager import GitManager
+
+__all__ = [
+    "GitBranchAnalyzer",
+    "GitManager",
+]
