+++ src/vibe_heal/vcs/git/manager.py
@@ -1,0 +1,347 @@
+"""Git operations manager."""
+
+from pathlib import Path
+
+import git
+
+from vibe_heal.ai_tools.base import AIToolType
+from vibe_heal.sonarqube.models import SonarQubeIssue, SonarQubeRule
+from vibe_heal.vcs.base import VCSManager
+from vibe_heal.vcs.exceptions import (
+    DirtyWorkingDirectoryError,
+    NotARepositoryError,
+    VCSOperationError,
+)
+
+
+class GitManager(VCSManager):
+    """Manages Git operations for vibe-heal."""
+
+    def __init__(self, repo_path: str | Path | None = None) -> None:
+        """Initialize Git manager.
+
+        Args:
+            repo_path: Path to Git repository (default: current directory)
+
+        Raises:
+            NotARepositoryError: If path is not a Git repository
+        """
+        self.repo_path = Path(repo_path or Path.cwd())
+
+        try:
+            self.repo = git.Repo(self.repo_path, search_parent_directories=True)
+        except git.InvalidGitRepositoryError as e:
+            msg = f"Not a Git repository: {self.repo_path}"
+            raise NotARepositoryError(msg) from e
+        except git.GitError as e:
+            msg = f"Git error: {e}"
+            raise VCSOperationError(msg) from e
+
+    def is_repository(self) -> bool:
+        """Check if current directory is a Git repository.
+
+        Returns:
+            True if in a Git repository
+        """
+        try:
+            return self.repo.git_dir is not None
+        except Exception:
+            return False
+
+    def is_clean(self) -> bool:
+        """Check if working directory is clean (no uncommitted changes).
+
+        Returns:
+            True if working directory is clean
+        """
+        return not self.repo.is_dirty(untracked_files=True)
+
+    def get_current_branch(self) -> str:
+        """Get the name of the current branch.
+
+        Returns:
+            Current branch name
+
+        Raises:
+            VCSOperationError: If unable to determine branch
+        """
+        try:
+            return self.repo.active_branch.name
+        except Exception as e:
+            msg = f"Unable to get current branch: {e}"
+            raise VCSOperationError(msg) from e
+
+    def get_uncommitted_files(self) -> list[str]:
+        """Get list of uncommitted files.
+
+        Returns:
+            List of file paths with uncommitted changes
+        """
+        changed = [item.a_path for item in self.repo.index.diff(None) if item.a_path is not None]
+        staged = [item.a_path for item in self.repo.index.diff("HEAD") if item.a_path is not None]
+        untracked = self.repo.untracked_files
+
+        return list(set(changed + staged + untracked))
+
+    def get_modified_or_staged_files(self) -> list[str]:
+        """Get list of modified or staged files (excluding untracked files).
+
+        Returns:
+            List of file paths that are modified or staged (but not untracked)
+        """
+        changed = [item.a_path for item in self.repo.index.diff(None) if item.a_path is not None]
+        staged = [item.a_path for item in self.repo.index.diff("HEAD") if item.a_path is not None]
+
+        return list(set(changed + staged))
+
+    def has_modified_or_staged_files(self) -> bool:
+        """Check if there are any modified or staged files (excluding untracked).
+
+        Returns:
+            True if there are modified or staged files
+        """
+        return len(self.get_modified_or_staged_files()) > 0
+
+    def get_all_modified_files(self) -> list[str]:
+        """Get all files that have been modified in the working directory.
+
+        This includes both staged and unstaged changes, but excludes untracked files.
+
+        Returns:
+            List of file paths with modifications
+        """
+        return self.get_modified_or_staged_files()
+
+    def has_uncommitted_changes(self, file_path: str) -> bool:
+        """Check if a specific file has uncommitted changes.
+
+        Args:
+            file_path: Path to file to check
+
+        Returns:
+            True if file has uncommitted changes or is staged
+        """
+        # Normalize path to match git's format
+        file_path_normalized = str(Path(file_path))
+
+        # Check if file is in changed files (working directory changes)
+        changed = [item.a_path for item in self.repo.index.diff(None)]
+        if file_path_normalized in changed:
+            return True
+
+        # Check if file is staged
+        staged = [item.a_path for item in self.repo.index.diff("HEAD")]
+        return file_path_normalized in staged
+
+    def commit_fix(
+        self,
+        issue: SonarQubeIssue,
+        files: list[str] | None,
+        ai_tool_type: AIToolType,
+        rule: SonarQubeRule | None = None,
+    ) -> str | None:
+        """Create a commit for a fixed issue.
+
+        If files is None, automatically detects and commits all modified files.
+        If files is an empty list, raises an error.
+
+        Args:
+            issue: The SonarQube issue that was fixed
+            files: List of files to commit, or None to auto-detect all modified files
+            ai_tool_type: The AI tool used for the fix
+            rule: Detailed rule information (optional)
+
+        Returns:
+            Commit SHA, or None if there were no changes to commit
+
+        Raises:
+            VCSOperationError: If commit fails or no files to commit
+        """
+        # Handle empty list case (explicit request with no files)
+        if files is not None and not files:
+            msg = "No files to commit"
+            raise VCSOperationError(msg)
+
+        # Auto-detect modified files if not provided
+        auto_detect = files is None
+        if auto_detect:
+            files = self.get_all_modified_files()
+
+        # If no modified files in auto-detect mode, check if already fixed
+        # This handles the case where one fix resolved multiple issues
+        if auto_detect and not files and not self.repo.index.diff("HEAD"):
+            # No modified files and nothing staged - issue was already fixed
+            return None
+        # If there are staged changes, proceed with those
+        # (This shouldn't normally happen in auto-detect mode, but handle it gracefully)
+
+        # At this point, files must be a list (either from auto-detect or passed in)
+        # This should never happen due to the logic above, but satisfy type checker
+        if files is None:
+            msg = "Internal error: files should not be None at this point"
+            raise VCSOperationError(msg)
+
+        # Determine correct file count for commit message
+        if auto_detect and not files:
+            staged_files_count = len(self.repo.index.diff("HEAD"))
+            file_count_for_message = staged_files_count
+        else:
+            file_count_for_message = len(files)
+        # Create commit message
+        message = self._create_commit_message(issue, ai_tool_type, rule, file_count_for_message)
+
+        # Stage and commit using helper method
+        return self._stage_and_commit(files, message)
+
+    def create_commit(
+        self,
+        message: str,
+        files: list[str] | None = None,
+        include_untracked: bool = False,
+    ) -> str | None:
+        """Create a commit with a custom message.
+
+        If files is None, automatically detects and commits all modified files.
+
+        Args:
+            message: Commit message
+            files: List of files to commit, or None to auto-detect all modified files
+            include_untracked: If True, include untracked (new) files in auto-detection
+
+        Returns:
+            Commit SHA, or None if there were no changes to commit
+
+        Raises:
+            VCSOperationError: If commit fails or no files to commit
+        """
+        # Auto-detect modified files if not provided
+        if files is None:
+            files = self.get_all_modified_files()
+            # Add untracked files if requested (e.g., for deduplication when AI creates new files)
+            if include_untracked:
+                untracked = self.repo.untracked_files
+                files.extend(untracked)
+
+        if not files:
+            msg = "No files to commit"
+            raise VCSOperationError(msg)
+
+        # Stage and commit using helper method
+        return self._stage_and_commit(files, message)
+
+    def _stage_and_commit(self, files: list[str], message: str) -> str | None:
+        """Stage files and create a commit if there are changes.
+
+        Args:
+            files: List of files to stage
+            message: Commit message
+
+        Returns:
+            Commit SHA, or None if there were no changes to commit
+
+        Raises:
+            VCSOperationError: If commit fails
+        """
+        try:
+            # Stage files (if any)
+            if files:
+                self.repo.index.add(files)
+
+            # Check if there are any changes to commit after staging
+            # This happens when one fix resolves multiple issues
+            if not self.repo.index.diff("HEAD"):
+                # No changes staged - return None to indicate no commit was created
+                return None
+
+            # Create commit
+            commit = self.repo.index.commit(message)
+
+            return commit.hexsha
+
+        except git.GitError as e:
+            msg = f"Failed to create commit: {e}"
+            raise VCSOperationError(msg) from e
+
+    def _create_commit_message(
+        self,
+        issue: SonarQubeIssue,
+        ai_tool_type: AIToolType,
+        rule: SonarQubeRule | None = None,
+        file_count: int = 1,
+    ) -> str:
+        """Create a formatted commit message for a fix.
+
+        Args:
+            issue: The SonarQube issue that was fixed
+            ai_tool_type: The AI tool used for the fix
+            rule: Detailed rule information (optional)
+            file_count: Number of files modified in this fix (default: 1)
+
+        Returns:
+            Formatted commit message
+        """
+        # Extract rule name (e.g., "python:S1481" -> "S1481")
+        rule_short = issue.rule.split(":")[-1] if ":" in issue.rule else issue.rule
+
+        # Create subject line (use rule name if available)
+        if rule:
+            subject = f"fix: [{issue.rule}] {issue.message[:50]}"
+        else:
+            subject = f"fix: [SQ-{rule_short}] {issue.message[:50]}"
+
+        if len(issue.message) > 50:
+            subject = subject.rstrip() + "..."
+
+        # Create body
+        body_parts = [
+            f"SonarQube Issue: {issue.key}",
+        ]
+
+        if rule:
+            body_parts.append(f"Rule: {issue.rule} - {rule.name}")
+        else:
+            body_parts.append(f"Rule: {issue.rule}")
+
+        body_parts.extend([
+            f"Severity: {issue.severity}",
+            f"Location: {issue.component}:{issue.line}",
+        ])
+
+        # Add file count if multiple files were modified
+        if file_count > 1:
+            body_parts.append(f"Files modified: {file_count}")
+
+        body_parts.append("")
+
+        # Add public documentation link
+        if rule:
+            body_parts.extend([
+                f"Rationale: {rule.public_doc_url}",
+                "",
+            ])
+
+        body_parts.append(f"Fixed by: vibe-heal using {ai_tool_type.display_name}")
+
+        body = "\n".join(body_parts)
+
+        # Combine subject and body
+        return f"{subject}\n\n{body}"
+
+    def require_clean_working_directory(self) -> None:
+        """Ensure working directory has no modified or staged files.
+
+        Untracked files are allowed, but modified or staged files must be committed first.
+
+        Raises:
+            DirtyWorkingDirectoryError: If there are modified or staged files
+        """
+        modified_or_staged = self.get_modified_or_staged_files()
+        if modified_or_staged:
+            msg = (
+                "Working directory has uncommitted changes. Please commit or stash modified files first.\n"
+                "Modified or staged files:\n"
+                + "\n".join(f"  - {f}" for f in modified_or_staged[:10])
+                + ("\n  ..." if len(modified_or_staged) > 10 else "")
+                + "\n\nNote: Untracked files are OK and won't block processing."
+            )
+            raise DirtyWorkingDirectoryError(msg)
